Add old script that unintentionally resizes to US
This was one of the earliest iterations from Grok. It was intended to
resize the PDF to be printed on B4, but all it seemed to do was resize
the PDF to be 8.5x11 inches. Re-adding this in order to use it for that
purpose.

The main script expects input to be 8.5x11 inches, so this can help
prepare documents for that if needed.